feat(analytics-docs): Optimize JSON output and improve component structure
- Updated buildData script to allow compact JSON output for reduced file size, with an option for pretty-printing during development.
- Refactored App component to utilize imported layout components, removing unnecessary styled components.
- Enhanced AttributesTable and AttributesTableRow components to improve prop naming for clarity.
- Updated Filter component to import Sort type from types for better type management.
- Streamlined useFilteredEvents by integrating utility functions for version comparison and event retrieval.
