+++ packages/analytics-docs/src/App.tsx
@@ -1,7 +1,5 @@
 import { useMemo } from 'react';
 
-import styled from 'styled-components';
-
 import {
     Box,
     Button,
@@ -12,35 +10,13 @@ import {
     Tooltip,
     intermediaryTheme,
 } from '@trezor/components';
-import { hexToRgba } from '@trezor/utils';
 
 import { EventCard } from './components/EventCard';
 import { Filter } from './components/Filter';
 import { GlobalStyle } from './components/GlobalStyle';
 import { ResultsInfo } from './components/ResultsInfo';
+import { Content, ContentContainer, TopBar } from './components/layout/AppLayout';
 import { useFilteredEvents } from './utils/useFilteredEvents';
-
-const TopBar = styled.div`
-    gap: 12px;
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 100;
-    background: ${({ theme }) => hexToRgba(theme.backgroundSurfaceElevation0, 0.8)};
-    box-shadow: ${({ theme }) => theme.boxShadowBase};
-    backdrop-filter: blur(20px);
-    padding: 12px 24px;
-`;
-
-const Content = styled.div`
-    margin: 140px 20px 20px;
-`;
-const ContentContainer = styled.div`
-    margin: auto;
-    max-width: 1000px;
-    width: 100%;
-`;
 
 export function App() {
     const {
