+++ packages/analytics-docs/src/utils/filterUtils.ts
@@ -1,0 +1,67 @@
+import type { EventDoc } from '../types';
+
+type AnalyticsJsonShape = {
+    events?: Record<string, EventDoc>;
+};
+
+export function getEventsFromJson(input: unknown): EventDoc[] {
+    const data = input as AnalyticsJsonShape;
+
+    if (!data || typeof data !== 'object') return [];
+    if (!data.events || typeof data.events !== 'object') return [];
+
+    return Object.values(data.events);
+}
+
+function parseVer(v?: string): readonly [number, number, number] {
+    const [a = 0, b = 0, c = 0] = (v ?? '').split('.').map(n => Number(n) || 0);
+
+    return [a, b, c];
+}
+
+function isSubsequence(word: string, segment: string): boolean {
+    let qi = 0;
+    for (let hi = 0; hi < segment.length && qi < word.length; hi++) {
+        if (segment[hi] === word[qi]) qi++;
+    }
+
+    return qi === word.length;
+}
+
+export function fuzzyMatch(query: string, haystack: string): boolean {
+    const tokens = query
+        .toLowerCase()
+        .trim()
+        .split(/[\s_\-/]+/)
+        .filter(Boolean);
+
+    if (tokens.length === 0) return true;
+
+    const segments = haystack.toLowerCase().split(/[/_]/);
+
+    for (const token of tokens) {
+        const found = segments.some(seg => isSubsequence(token, seg));
+        if (!found) return false;
+    }
+
+    return true;
+}
+
+export function compareVersionsDesc(va?: string, vb?: string): number {
+    const [a1, b1, c1] = parseVer(va);
+    const [a2, b2, c2] = parseVer(vb);
+
+    if (a1 !== a2) return a2 - a1;
+    if (b1 !== b2) return b2 - b1;
+    if (c1 !== c2) return c2 - c1;
+
+    return 0;
+}
+
+export function getEventAddedVersion(e: EventDoc): string | undefined {
+    return e.changelog?.addedInVersion;
+}
+
+export function getEventUpdatedVersion(e: EventDoc): string | undefined {
+    return e.changelog?.lastUpdatedInVersion ?? e.changelog?.addedInVersion;
+}
